(MarsProductFilter) fix ext filter

+++ Module.Frontend.TPM/Client/model/tpm/product/Product.js
@@ -39,7 +39,7 @@
 		{ name: 'UOM_PC2Case', useNull: true, type: 'int', hidden: false, isDefault: true },
         { name: 'Division', useNull: true, type: 'int', hidden: false, isDefault: true },
         { name: 'UOM', useNull: true, type: 'string', hidden: false, isDefault: true },
-        { name: 'NetWeight', useNull: true, type: 'double', hidden: false, isDefault: true },
+        { name: 'NetWeight', useNull: true, type: 'float', hidden: false, isDefault: true },
         { name: 'Checked', useNull: false, type: 'bool', hidden: true, isDefault: false, defaultValue: false }
     ],
     proxy: {
